refactored AdjustableMinecraftClock to a new file
